feat(InvoiceManagement): Introduce invoice filtering and details components
- Add useInvoiceFilter hook for managing invoice search and status filtering.
- Create FilteredInvoiceList component to display filtered invoices with a user-friendly interface.
- Refactor InvoicePage to integrate new invoice filtering functionality and improve UI consistency.
- Implement loading, error, and empty states for invoice views to enhance user experience.
- Update routing to support new invoice details structure and improve navigation.
