+++ src/page/store/account/index.ts
@@ -2,7 +2,7 @@ export { default as AddressesPage } from "./addresses";
 export { default as CartPage } from "./cart";
 export { default as ChangePasswordPage } from "./change-password";
 export { default as InvoicePage } from "./invoice";
-export { default as InvoiceDetailsPage } from "./invoice/details";
+export { default as InvoiceDetailsPage } from "./invoice/[id]/page";
 export { default as OrderPage } from "./order";
 export { default as RootAccountPage } from "./root";
