+++ src/page/store/account/order/[id]/[id].details-skeleton.tsx
