+++ src/page/store/account/order/[id]/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./page";
