+++ src/page/store/account/order/index.ts
@@ -1,2 +1,2 @@
 export { default } from "./page";
-export { default as OrderDetails } from "./order.details";
+export { default as OrderDetails } from "./[id]/page";
